+++ src/jsx/components/Footer.js
@@ -1,11 +1,12 @@
 import React, { Component } from 'react'
+import {version} from '../../../package.json'
 
 class Footer extends Component {
     render() {
         return (
             <footer className="box">
-                <h2>Dakara Project</h2>
-                <p className="version">Proof of Concept</p>
+                <h2>Dakara client</h2>
+                <p className="version">{version}</p>
             </footer>
         )
     }