+++ CHANGELOG.md
@@ -33,3 +33,4 @@
 ### Added
 
 - Changelog.
+- Display version in footer.